So I looked into the specs that were taking a long time to run, and most of them came to the same problem, where we were using capybara to fill in textareas where summernote was present. This is slow cause capybara enters character by character, which fired callbacks for each change in character in summernote (learnt this from @jeremyyap 👍 ).

This PR changes those lines to use a helper which speeds tests up quickly. I'm not sure how much of an impact that is on our CI since we have parallelised tests. I've also modified the helper a little as the implementation of rails and react summernote is slightly different. 

Timings for the tests which were fixed: 
```
# Before
Finished in 3 minutes 1.4 seconds (files took 6.35 seconds to load)
5 examples, 0 failures
# After
Finished in 27.59 seconds (files took 5.33 seconds to load)
5 examples, 0 failures
```