@allenwq 2017-03-23
@weiqingtoh So our the orginal `fill_in_summernote` is also slow ?