@weiqingtoh 2017-03-23
@allenwq `fill_in_summernote` is fast, I had to refactor it cause the selector for the rails and react implementation is slightly different so I generalised it.